requirements.txt
appdirs==1.4.4
asgiref==3.2.10
boto3==1.16.60
botocore==1.19.60
certifi==2020.6.20
distlib==0.3.1
Django==3.1.1
django-filter==2.4.0
django-storages==1.11.1
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.7.0
jmespath==0.10.0
Pillow==8.1.0
pipenv==2020.8.13
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.3
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0
xxx==0.0.1
zipp==3.1.0
